% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.leaflet.plot.R
\name{import.palettes}
\alias{import.palettes}
\title{Import some user-defined palettes}
\usage{
import.palettes()
}
\description{
The function imports some user-defined colour palettes.
}
\examples{
\dontrun{
import.palettes()
}
}
