% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgm.R
\docType{data}
\name{lgm}
\alias{lgm}
\title{Example dataset}
\format{
A data frame containing species name, locality ID, occurrence longitude, occurrence latitude and averaged age estimation.
}
\usage{
data(lgm)
}
\description{
The dataset including occurrences of large mammals lived in Eurasia during the Last Glacial Maximum, along with average age estimations and locality ID.
}
\references{
Carotenuto, F., Barbera, C., & Raia, P. (2010). Occupancy, range size, and phylogeny in Eurasian Pliocene to Recent large mammals. Paleobiology, 36(3), 399-414.

Carotenuto, F., Diniz-Filho, J. A. F., & Raia, P. (2015). Space and time: The two dimensions of Artiodactyla body mass evolution. Palaeogeography, Palaeoclimatology, Palaeoecology, 437, 18-25.

Carotenuto, F., Di Febbraro, M., Melchionna, M., Castiglione, S., Saggese, F., Serio, C., … & Raia, P. (2016). The influence of climate on species distribution over time and space during the late Quaternary. Quaternary Science Reviews, 149, 188-199.

Carotenuto, F., Di Febbraro, M., Melchionna, M., Mondanaro, A., Castiglione, S., Serio, C., ... & Raia, P. (2018). The well-behaved killer: Late Pleistocene humans in Eurasia were significantly associated with living megafauna only. Palaeogeography, palaeoclimatology, palaeoecology, 500, 24-32.
}
