% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapofpast.R
\name{mapofpast}
\alias{mapofpast}
\title{Creates prediction ground and paleo dem}
\usage{
mapofpast(Ma,domain,ext=NULL,projection,pts=NULL)
}
\arguments{
\item{Ma}{An integer value comprised in between 0 and 540}

\item{domain}{Character. Eithter \code{"land"}, for terrestrial species, or \code{"sea"}, for marine species.}

\item{ext}{Optional. A numeric vector with the coordinates (in wgs84 format) of the rectangle's vertices enclosing the domain. It must be c(min longitude, max longitude, min latitude, max latitude). See details}

\item{projection}{Character. This is the euqual-area projection of the final maps. A CRS object in the proj4 format or either \code{"moll"} (Mollweide) or \code{"laea"} (Lambert Azimuthal equal area) projections. See details.}

\item{pts}{Optional. A dataframe with the x and y coordinates of individual fossil localities. If specified the mean center of the fossil record spatial distribution is considered as projection's centre. See details.}
}
\value{
A list of two rasters to be used in the \code{\link{minosse.data}} function: the first is the prediction domain to be used in the \code{prediction.ground} argument, and the second is a paleodem to be used in the \code{abiotic.covs} argument.
}
\description{
This function uses Scotese and Wright (2018) PALEOMAPS to generate the prediction ground and paleo dems of the requested temporal interal
}
\details{
By setting a numeric vector for the argumento \code{ext}, the user can select a portion of the choosen domain.
If the argument is NULL, when running the function, the user is asket to draw the polygon encolosing the portion
of the domain to be used in all the predictions. In this case, the user will draw a polygon's vertices by clicking
on the map with the lef button of the mouse. Then, by pressing Esc, the function will crop the portion of the domain
enclosed by the drawn polygon. If the argument \code{projection} is either \code{"laea"} or \code{"moll"}, the projection's centre is computed
as the extent's centre either if set in the \code{ext} argument either if set by drawing a polygon. In the case the argument \code{pts}
is specified, then the mean centre of the fossil record distribution is used as projection centre.
}
\examples{
  \donttest{
  mapofpast(Ma=540,domain="land",ext=c(-50,50,-30,30),projection="laea",pts=NULL)
  }
}
\author{
Francesco Carotenuto, francesco.carotenuto@unina.it
}
