% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minosse.poly.R
\name{minosse.poly}
\alias{minosse.poly}
\title{The species geographic range polygon}
\usage{
minosse.poly(minosse_res, th_num=3)
}
\arguments{
\item{minosse_res}{the output of \code{\link{minosse.target}} function or the \code{minosse_res} element list in the minosse function output.}

\item{th_num}{Numeric. The Regression Kriging prediction map binarization threshold value index as reported in the \code{optimal.thresholds} function in the PresenceAbsence package. Default 3 = MaxSens+Spec,    maximizes (sensitivity+specificity)/2.}
}
\value{
The target species geographic range as a SpatialPolygonsDataFrame object.
}
\description{
This function turns \code{\link{minosse}} geographic range raster map into a spatial polygon with the related surface area.
}
\author{
Francesco Carotenuto, francesco.carotenuto@unina.it
}
