% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minosse.target.R
\name{minosse.target}
\alias{minosse.target}
\title{The target species geographic range computation.}
\usage{
minosse.target(resp,predictors,bkg="presence",min.bkg=NULL,n.sims=10,
n.folds=1,sampling.by.distance=TRUE,n.sims.clusters=NULL,seed=NULL)
}
\arguments{
\item{resp}{The target species occurrence record. Usually the first element list in the \code{\link{minosse.data}} output.}

\item{predictors}{The predictor species raster stack. Usually the second element list in the \code{minosse.data} output.}

\item{bkg}{The number of pseudo absences to be simulated for each predictor species. If \code{"presence"}, \code{minosse.data} generates as many pseudo absences the presences in each species.}

\item{min.bkg}{Numeric. If \code{bkg.predictors} is set to \code{"presence"}, this is the minimum number of pseudo absences to simulate if a species occurrence number is below this value.}

\item{n.sims}{Numeric. The number of pseudo absences simulations (see details).}

\item{n.folds}{Numeric. The number of folds for AUC-based cross-validation of Regression Kriging predictions. Default 1.}

\item{sampling.by.distance}{Logical. If \code{TRUE} pseudo absences are simulated with an intensity proportional to the distance to the presence data. If \code{FALSE} a pure spatial random distribution is simulated.}

\item{n.sims.clusters}{Numeric or \code{NULL}. The number of machine cores to use when setting multiple pseudo absences simulations. Default \code{"automatic"}. Default is \code{NULL}.}

\item{seed}{Numeric. The \code{seed} to use for experiment replication.}
}
\value{
A list of 7 objects is returned. The first item of the list is a list of 12 geographic range maps obtained by binarizing Regression Kriging output by means of an equivalent number of threshold values. The second element
of the list is the probability map yielded by Regression Kriging. The third element is a list containing eXtreme Gradient Boosting model and validation statistics for each of the replicates of the pseudo absences simulations (\code{n.sims}, see above).
The fourth element of the list shows the Regression Kriging binarization threshold values. The fifth element (validation) shows the Regression Kriging prediction (RMSE and total explained variance) and xgboost model validations (squared R and RMSE).
The two last elements show the target species name and the occurence record.
}
\description{
This function reconstructs the target species geographic range and probability distribution raster maps.
}
\details{
For Machine Learning Algorithms Barbet & Masin (https://doi.org/10.1111/j.2041-210X.2011.00172.x) suggest using as many pseudo absences as the number of the presences. For very poor occurrence datasets
(with less than 1000 presences), they suggest performing multiple (10 at least) pseudo absences simulations and then averaging the derived predictions. \code{minosse.target} function allows fully complying with these authors' suggestions.
}
\examples{
  \dontrun{
  library(raster)
  data(lgm)
  raster(system.file("exdata/prediction_ground.gri", package="EcoPast"))->prediction_ground

  minosse_dat<-minosse.data(obj=lgm,species_name="Mammuthus_primigenius",
  domain=NULL,coc.by="locality",min.occs=3,abiotic.covs=NULL,
  combine.covs=FALSE,reduce_covs_by="pca",covs_th=0.95,c.size="mean",
  bkg.predictors="presence",min.bkg=100,sampling.by.distance=TRUE,
  crop.by.mcp=FALSE,prediction.ground=prediction_ground,projection=NULL,
  lon_0=NULL,lat_0=NULL,n.clusters="automatic",seed=625)

  minosse_res<-minosse.target(resp=minosse_dat[[1]],predictors=minosse_dat[[2]],
  bkg="presence", min.bkg = 100,n.sims=10,n.folds=1,sampling.by.distance=TRUE,
  n.sims.clusters=NULL,seed=625)

  }
}
\author{
Francesco Carotenuto, francesco.carotenuto@unina.it
}
