% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.desaggregation.RASTER.R
\name{occ.desaggregation.RASTER}
\alias{occ.desaggregation.RASTER}
\title{Remove duplicated points falling into the same raster cell}
\usage{
occ.desaggregation.RASTER(df, colxy, rast, plot=TRUE)
}
\arguments{
\item{df}{a matrix or data.frame containing the points coordinates.}

\item{colxy}{a numeric vector indicating the IDs of the x-y columns.}

\item{rast}{a \code{RasterLayer} object to be used as base raster map.}

\item{plot}{a logical vector indicating if a plot must be displayed.}
}
\value{
A matrix including the reduced points
}
\description{
The function removes duplicated points falling into the same raster cell.
}
\examples{
\donttest{
library(raster)
raster(system.file("exdata/prediction_ground.gri", package="EcoPast"))->prediction_ground
data(lgm)

data_mammuth<-subset(lgm, spec=="Mammuthus_primigenius")

coordinates(data_mammuth)<-~x+y
proj4string(data_mammuth)<-"+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"
data_mammuth<-spTransform(data_mammuth, proj4string(prediction_ground))
data_mammuth<-as.data.frame(data_mammuth)

data_mammuth_reduced<-occ.desaggregation.RASTER(df=data_mammuth, colxy=4:5,
rast=prediction_ground, plot=TRUE)

}
}
\author{
Mirko Di Febbraro
}
