% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.leaflet.plot.R
\name{raster.leaflet.plot}
\alias{raster.leaflet.plot}
\title{Plot a \code{RasterLayer} or \code{RasterStack} object through the leaflet functionalities}
\usage{
raster.leaflet.plot(RASTER.LAYER,PROJECT=FALSE,DOMAIN=NULL,PALETTE=NULL,
POSITION=NULL,REVERSE=TRUE,OPACITY=1,colNA=NA)
}
\arguments{
\item{RASTER.LAYER}{a \code{RasterLayer} or \code{RasterStack} object to be plotted.}

\item{PROJECT}{a logical vector indicating if the raster map must be projected to EPSG:3857.}

\item{DOMAIN}{a numerical vector indicating the range of raster values to be plotted.}

\item{PALETTE}{one of the palettes from PAL1 to PAL28.}

\item{POSITION}{a character vector indicating legend position ('topright' is the default).}

\item{REVERSE}{a logical vector indicating if to reverse the color palette.}

\item{OPACITY}{a numerical vector ranging from 0 to 1 indicating the opacity level.}

\item{colNA}{the color of NA cells.}
}
\value{
A data.frame with the reduced points.
}
\description{
The function uses the \code{leaflet} package to plot raster maps, also allowing zoom and navigation functionalities.
}
\examples{
\donttest{
  library(raster)
  data(lgm)
  raster(system.file("exdata/prediction_ground.gri",
         package="EcoPast"))->prediction_ground
  raster.leaflet.plot(prediction_ground)
}
}
\author{
Mirko Di Febbraro
}
