% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.NAs.stack.R
\name{remove.NAs.stack}
\alias{remove.NAs.stack}
\title{Remove unpaired NA cells from a RasterStack}
\usage{
remove.NAs.stack(rast.stack)
}
\arguments{
\item{rast.stack}{a \code{RasterStack} object where to remove NAs.}
}
\value{
A \code{RasterStack} object without unpaired NA cells.
}
\description{
The function removes unpaired NA cells from a RasterStack.
}
\examples{
\donttest{
library(raster)
raster(system.file("exdata/prediction_ground_thinned.gri",
package="EcoPast"))->prediction_ground_thinned
raster(system.file("exdata/prediction_ground.gri", package="EcoPast"))->prediction_ground
s1<-stack(prediction_ground, prediction_ground_thinned)
plot(s1)

summary(rasterToPoints(s1))

s2<-remove.NAs.stack(s1)

summary(rasterToPoints(s2))
}
}
\author{
Mirko Di Febbraro
}
