% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.leaflet.plot.R
\name{show.custom.palettes}
\alias{show.custom.palettes}
\title{Show the palettes from PAL1 to PAL28}
\usage{
show.custom.palettes()
}
\description{
The function shows an example of the colour palettes imported through \code{import.palettes()}
}
\examples{
\dontrun{
show.custom.palettes()
}
}
